factor WellKnownTypeFilterHook into a new filter package
