+++ internal/servers/controller/handlers/filtering.go
@@ -1,13 +1,9 @@
 package handlers
 
 import (
-	"reflect"
-
 	"github.com/hashicorp/boundary/internal/errors"
+	"github.com/hashicorp/boundary/internal/filter"
 	"github.com/hashicorp/go-bexpr"
-	"google.golang.org/protobuf/types/known/structpb"
-	"google.golang.org/protobuf/types/known/timestamppb"
-	"google.golang.org/protobuf/types/known/wrapperspb"
 )
 
 // filterItem captures all the different namespaces that can be used when
@@ -27,7 +23,7 @@ func NewFilter(f string) (*Filter, error) {
 	if f == "" {
 		return &Filter{}, nil
 	}
-	e, err := bexpr.CreateEvaluator(f, bexpr.WithTagName("json"), bexpr.WithHookFn(wellKnownTypeFilterHook))
+	e, err := bexpr.CreateEvaluator(f, bexpr.WithTagName("json"), bexpr.WithHookFn(filter.WellKnownTypeFilterHook))
 	if err != nil {
 		return nil, errors.Wrap(err, op, errors.WithMsg("couldn't build filter"), errors.WithCode(errors.InvalidParameter))
 	}
@@ -48,38 +44,4 @@ func (f *Filter) Match(item interface{}) bool {
 	// is part of a sub structure that is nil in this item. Because of this, any filter which would
 	// result in an error using the underlying library is simply interpreted as not a match.
 	return err == nil && m
-}
-
-// wellKnownTypeFilterHook is passed to bexpr to treat all proto well-known
-// types as the types they wrap for comparison.
-func wellKnownTypeFilterHook(v reflect.Value) reflect.Value {
-	if !v.CanInterface() {
-		return v
-	}
-	ret := v.Interface()
-	switch pm := v.Interface().(type) {
-	case *wrapperspb.BoolValue:
-		ret = pm.GetValue()
-	case *wrapperspb.BytesValue:
-		ret = pm.GetValue()
-	case *wrapperspb.StringValue:
-		ret = pm.GetValue()
-	case *wrapperspb.DoubleValue:
-		ret = pm.GetValue()
-	case *wrapperspb.FloatValue:
-		ret = pm.GetValue()
-	case *wrapperspb.Int32Value:
-		ret = pm.GetValue()
-	case *wrapperspb.Int64Value:
-		ret = pm.GetValue()
-	case *wrapperspb.UInt32Value:
-		ret = pm.GetValue()
-	case *wrapperspb.UInt64Value:
-		ret = pm.GetValue()
-	case *structpb.Struct:
-		ret = pm.AsMap()
-	case *timestamppb.Timestamp:
-		ret = pm.AsTime()
-	}
-	return reflect.ValueOf(ret)
 }
