+++ internal/servers/controller/handlers/filtering_test.go
@@ -1,7 +1,6 @@
 package handlers
 
 import (
-	"reflect"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -116,22 +115,5 @@ func TestNewFilter(t *testing.T) {
 			require.NoError(t, err)
 			assert.Equal(t, tc.match, f.Match(tc.in))
 		})
-	}
-}
-
-func TestWellKnownTypeFilterHook(t *testing.T) {
-	conversions := map[interface{}]interface{}{
-		wrapperspb.String("foo"):        "foo",
-		wrapperspb.UInt64(123):          uint64(123),
-		wrapperspb.Int64(123):           int64(123),
-		wrapperspb.UInt32(123):          uint32(123),
-		wrapperspb.Int32(123):           int32(123),
-		wrapperspb.Float(123):           float32(123),
-		wrapperspb.Double(123):          float64(123),
-		wrapperspb.Bytes([]byte("foo")): []byte("foo"),
-		wrapperspb.Bool(true):           true,
-	}
-	for in, out := range conversions {
-		assert.Equal(t, reflect.ValueOf(out).Interface(), wellKnownTypeFilterHook(reflect.ValueOf(in)).Interface())
 	}
 }
